+++ src/middleware.ts
@@ -1,10 +1,90 @@
-import { NextResponse, NextRequest } from 'next/server'
- import {updateSession} from "@/utils/supabase/middleware";
+import { createServerClient, type CookieOptions } from "@supabase/ssr";
+import { NextResponse, type NextRequest } from "next/server";
 
- export async function middleware(request: NextRequest ) {
-    return await updateSession(request);   
- }
- 
+export async function middleware(request: NextRequest) {
+  let response = NextResponse.next({
+    request: {
+      headers: request.headers,
+    },
+  });
+
+  const path = new URL(request.url).pathname;
+
+  const protectedRoutes = ["/admin"];
+  const authRoutes = ["/login"];
+
+  const isProtectedRoute = protectedRoutes.some((route) => path.startsWith(route));
+  const isAuthRoute = authRoutes.includes(path);
+
+  if (isProtectedRoute || isAuthRoute) {
+    const user = await getUser(request, response);
+
+    if (isProtectedRoute && !user) {
+      return NextResponse.redirect(new URL("/login", request.url));
+    }
+
+    if (isAuthRoute && user) {
+      return NextResponse.redirect(new URL("/", request.url));
+    }
+  }
+
+  return response;
+}
+
 export const config = {
-  matcher: '/about/:path*',
+  matcher: [
+    "/((?!_next/static|_next/image|favicon.ico|.*\\.(?:svg|png|jpg|jpeg|gif|webp)$).*)",
+  ],
+};
+
+async function getUser(request: NextRequest, response: NextResponse) {
+  const supabaseClient = createServerClient(
+    process.env.NEXT_PUBLIC_SUPABASE_URL!,
+    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
+    {
+      cookies: {
+        get(name: string) {
+          return request.cookies.get(name)?.value;
+        },
+        set(name: string, value: string, options: CookieOptions) {
+          request.cookies.set({
+            name,
+            value,
+            ...options,
+          });
+          response = NextResponse.next({
+            request: {
+              headers: request.headers,
+            },
+          });
+          response.cookies.set({
+            name,
+            value,
+            ...options,
+          });
+        },
+        remove(name: string, options: CookieOptions) {
+          request.cookies.set({
+            name,
+            value: "",
+            ...options,
+          });
+          response = NextResponse.next({
+            request: {
+              headers: request.headers,
+            },
+          });
+          response.cookies.set({
+            name,
+            value: "",
+            ...options,
+          });
+        },
+      },
+    }
+  );
+
+  const user = (await supabaseClient.auth.getUser()).data.user;
+
+  return user;
 }
